requirements.txt
streamlit
transformers
faiss-cpu
openai
langchain
torch
requests
beautifulsoup4
huggingface_hub
langchain-cerebras
cerebras_cloud_sdk
sentence_transformers
